unify create and update pages
